Move stringify() & strOptions from schema/_string.js to stringify.js
