+++ schema.js
@@ -1,3 +1,3 @@
 module.exports = require('./dist/schema')
 module.exports.nullOptions = require('./dist/schema/core').nullOptions
-module.exports.strOptions = require('./dist/schema/_string').strOptions
+module.exports.strOptions = require('./dist/stringify').strOptions
